+++ src/test/resources/com/littleh322/springboot/springboot/TestRunner.java
@@ -4,8 +4,6 @@ import org.junit.runner.JUnitCore;
 import org.junit.runner.Result;
 import org.junit.runner.notification.Failure;
 
-import com.littleh322.springboot.springboot.tests.ParameterizedEmployees;
-
 public class TestRunner {
 	public static void main(String[] args) {
 		Result result = JUnitCore.runClasses(ParameterizedEmployees.class);
